pyproject.toml
[tool.poetry]
name = "info002-tp2"
version = "0.1.0"
description = ""
authors = ["Simon LEONARD <git-1001af4@sinux.sh>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pillow = "^10.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
